move additional transformer tests to new pkg from backend
