+++ core/transformer/src/test/standalone/RenderTimeline.test.ts
@@ -3,13 +3,13 @@
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
 import { expect } from "chai";
-import { Guid, Id64, Id64String, OpenMode } from "@bentley/bentleyjs-core";
+import { CompressedId64Set, Guid, Id64, Id64String, OpenMode } from "@bentley/bentleyjs-core";
 import { Box, Point3d, Vector3d, YawPitchRollAngles } from "@bentley/geometry-core";
 import {
   Code, GeometryStreamBuilder, IModel, PhysicalElementProps, RenderSchedule, RenderTimelineProps,
 } from "@bentley/imodeljs-common";
 import {
-  GenericSchema, IModelJsFs, PhysicalModel, PhysicalObject, PhysicalPartition, RenderTimeline, StandaloneDb, SubjectOwnsPartitionElements,
+  GenericSchema, IModelJsFs, IModelTransformer, PhysicalModel, PhysicalObject, PhysicalPartition, RenderTimeline, SpatialCategory, StandaloneDb, SubjectOwnsPartitionElements,
 } from "../../imodeljs-backend";
 import { IModelTestUtils } from "../IModelTestUtils";
 
@@ -125,5 +125,63 @@ describe("RenderTimeline", () => {
     expect(timeline.scriptProps).to.deep.equal(scriptProps);
 
     imodel.close();
+  });
+
+  it("remaps schedule script Ids on clone", async () => {
+    const sourceDb = createIModel("remap-source");
+    const model = insertPhysicalModel(sourceDb);
+    const category = SpatialCategory.insert(sourceDb, IModel.dictionaryId, "cat", {});
+    const elementIds: Id64String[] = [];
+    for (let i = 0; i < 3; i++)
+      elementIds.push(insertPhysicalElement(sourceDb, model, category));
+
+    const scriptProps: RenderSchedule.ScriptProps = [{
+      modelId: model,
+      elementTimelines: [{
+        batchId: 1,
+        elementIds,
+        visibilityTimeline: [{ time: 42, value: 50 }],
+      }],
+    }];
+
+    const sourceTimelineId = insertTimeline(sourceDb, scriptProps);
+    sourceDb.saveChanges();
+
+    // Make sure targetDb has more elements in it to start with than sourceDb does, so element Ids will be different.
+    const targetDb = createIModel("remap-target");
+    for (let i = 0; i < 3; i++)
+      insertPhysicalModel(targetDb);
+
+    targetDb.saveChanges();
+
+    const transformer = new IModelTransformer(sourceDb, targetDb);
+    await transformer.processAll();
+
+    const targetTimelineIds = targetDb.queryEntityIds({ from: RenderTimeline.classFullName });
+    expect(targetTimelineIds.size).to.equal(1);
+    let targetTimelineId: Id64String | undefined;
+    for (const id of targetTimelineIds)
+      targetTimelineId = id;
+
+    const sourceTimeline = sourceDb.elements.getElement<RenderTimeline>(sourceTimelineId);
+    expect(sourceTimeline.scriptProps).to.deep.equal(scriptProps);
+
+    const targetTimeline = targetDb.elements.getElement<RenderTimeline>(targetTimelineId!);
+    expect(targetTimeline.scriptProps).not.to.deep.equal(scriptProps);
+    expect(targetTimeline.scriptProps.length).to.equal(1);
+
+    expect(targetTimeline.scriptProps[0].modelId).not.to.equal(scriptProps[0].modelId);
+    expect(targetDb.models.getModel<PhysicalModel>(targetTimeline.scriptProps[0].modelId)).instanceof(PhysicalModel);
+
+    expect(targetTimeline.scriptProps[0].elementTimelines.length).to.equal(1);
+    const timeline = targetTimeline.scriptProps[0].elementTimelines[0];
+
+    let numElements = 0;
+    expect(typeof timeline.elementIds).to.equal("string"); // remapping also compresses Ids if they weren't already.
+    for (const elementId of CompressedId64Set.iterable(timeline.elementIds as string)) {
+      ++numElements;
+      expect(targetDb.elements.getElement<PhysicalObject>(elementId)).instanceof(PhysicalObject);
+    }
+    expect(numElements).to.equal(3);
   });
 });
